Add an explicit cast to prevent a compiler warning.
git-svn-id: svn://svn.tartarus.org/sgt/halibut@4399 cda61777-01e9-0310-a592-d414129be87e

+++ winhelp.c
@@ -418,7 +418,7 @@ static unsigned long context_hash(char *context)
 
             if (val > 0 && hash > (0xFFFFFFFFUL - val)) {
                 hash -= (0xFFFFFFFFUL - val) + 1;
-            } else if (val < 0 && hash < -val) {
+            } else if (val < 0 && hash < (unsigned long)-val) {
                 hash += (0xFFFFFFFFUL + val) + 1;
             } else
                 hash += val;
